refactor: move can wrappers to system_init
